+++ gdsfactory/components/taper_cross_section.py
@@ -7,7 +7,7 @@ from gdsfactory.cell import cell
 from gdsfactory.component import Component
 from gdsfactory.cross_section import rib_conformal, strip_rib_tip
 from gdsfactory.typings import CrossSectionSpec
-from gdsfactory.pathlengths import route_info
+from gdsfactory.route_info import route_info
 
 
 @cell
